# docs/swagger.yaml
definitions:
  ai_m.AiPromptData:
    properties:
      code:
        type: string
      name:
        properties:
          en:
            type: string
          zh:
            type: string
        type: object
      subTabs:
        items:
          properties:
            code:
              type: string
            name:
              properties:
                en:
                  type: string
                zh:
                  type: string
              type: object
            prompts:
              items:
                properties:
                  code:
                    type: string
                  en:
                    type: string
                  zh:
                    type: string
                type: object
              type: array
          type: object
        type: array
    type: object
...
info:
  contact: {}
  license: {}
paths:
  /v1_0/ai/prompts:
    get:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ai_m.AiPromptData'
            type: array
      summary: 获取所有提示词
      tags:
      - AIPrompts
...
  /v1_0/vap/vapc:
    post:
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: mp4
        in: formData
        name: mp4
        required: true
        type: string
      responses:
        "200": {}
      summary: 获取vapc
      tags:
      - VAP
swagger: "2.0"